Move SQR.importStatuses to TasksRepository.
